+++ packages/components/src/components/input/input.ts
@@ -7,6 +7,7 @@ import { HasSlotController } from '../../internal/slot';
 import { ifDefined } from 'lit/directives/if-defined.js';
 import { live } from 'lit/directives/live.js';
 import { LocalizeController } from '../../utilities/localize';
+import { longPress } from '../../internal/longpress.js';
 import { property, query, state } from 'lit/decorators.js';
 import { watch } from '../../internal/watch';
 import componentStyles from '../../styles/component.styles';
@@ -41,6 +42,7 @@ const isFirefox = isChromium ? false : navigator.userAgent.includes('Firefox');
  * @slot show-password-icon - An icon to use in lieu of the default show password icon.
  * @slot hide-password-icon - An icon to use in lieu of the default hide password icon.
  * @slot help-text - Text that describes how to use the input. Alternatively, you can use the `help-text` attribute.
+ * @slot tooltip - An optional tooltip that helps describe the input. Use this slot with the `sd-tooltip` component.
  *
  * @event sd-blur - Emitted when the control loses focus.
  * @event sd-change - Emitted when an alteration to the control's value is committed by the user.
@@ -146,8 +148,8 @@ export default class SdInput extends SolidElement implements SolidFormControl {
   /** Determines whether or not the password is currently visible. Only applies to password input types. */
   @property({ attribute: 'password-visible', type: Boolean }) passwordVisible = false;
 
-  /** Hides the browser's built-in increment/decrement spin buttons for number inputs. */
-  @property({ attribute: 'no-spin-buttons', type: Boolean }) noSpinButtons = false;
+  /** Hides the browser's built-in increment/decrement spin buttons for number inputs and displays custom buttons. */
+  @property({ attribute: 'spin-buttons', type: Boolean }) spinButtons = false;
 
   /** The minimum length of input that will be considered valid. */
   @property({ type: Number }) minlength: number;
@@ -316,6 +318,47 @@ export default class SdInput extends SolidElement implements SolidFormControl {
     this.passwordVisible = !this.passwordVisible;
   }
 
+  private isDecrementDisabled() {
+    if (this.disabled || this.readonly) {
+      return true;
+    }
+
+    if (this.min === undefined || this.min === null) {
+      return false;
+    }
+
+    const min = typeof this.min === 'string' ? parseFloat(this.min) : this.min;
+    return this.valueAsNumber <= min;
+  }
+
+  private isIncrementDisabled() {
+    if (this.disabled || this.readonly) {
+      return true;
+    }
+
+    if (this.max === undefined || this.max === null) {
+      return false;
+    }
+
+    const max = typeof this.max === 'string' ? parseFloat(this.max) : this.max;
+    return this.valueAsNumber >= max;
+  }
+
+  private handleStep() {
+    this.handleInput();
+    this.input.focus();
+  }
+
+  private handleStepUp() {
+    this.stepUp();
+    this.handleStep();
+  }
+
+  private handleStepDown() {
+    this.stepDown();
+    this.handleStep();
+  }
+
   @watch('disabled', { waitUntilFirstUpdate: true })
   handleDisabledChange() {
     // Disabled form controls are always valid
@@ -473,15 +516,19 @@ export default class SdInput extends SolidElement implements SolidFormControl {
     // Render
     return html`
       <div part="form-control" class=${cx(this.disabled && 'pointer-events-none')}>
-        <label
-          part="form-control-label"
-          id="label"
-          class=${cx('mb-2', hasLabel ? 'inline-block' : 'hidden', textSize)}
-          for="input"
-          aria-hidden=${hasLabel ? 'false' : 'true'}
-        >
-          <slot name="label">${this.label}</slot>
-        </label>
+        <div class="flex items-center gap-1 mb-2">
+          <label
+            part="form-control-label"
+            id="label"
+            class=${cx(hasLabel ? 'inline-block' : 'hidden', textSize)}
+            for="input"
+            aria-hidden=${hasLabel ? 'false' : 'true'}
+          >
+            <slot name="label">${this.label}</slot>
+          </label>
+
+          <slot name="tooltip"></slot>
+        </div>
 
         <div part="form-control-input" class="relative w-full">
           <div
@@ -536,7 +583,7 @@ export default class SdInput extends SolidElement implements SolidFormControl {
               pattern=${ifDefined(this.pattern)}
               enterkeyhint=${ifDefined(this.enterkeyhint)}
               inputmode=${ifDefined(this.inputmode)}
-              aria-describedby="help-text"
+              aria-describedby="help-text invalid-message"
               @change=${this.handleChange}
               @input=${this.handleInput}
               @invalid=${this.handleInvalid}
@@ -644,6 +691,49 @@ export default class SdInput extends SolidElement implements SolidFormControl {
                   class=${cx('inline-flex', iconColor, iconMarginLeft, iconSize)}
                 ></slot>`
               : ''}
+            ${this.type === 'number' && this.spinButtons
+              ? html`
+                  <div part="stepper" class="flex items-center">
+                    <button
+                      part="decrement-number-stepper"
+                      class="stepper-button flex"
+                      type="button"
+                      ?disabled=${this.isDecrementDisabled()}
+                      aria-hidden="true"
+                      ${longPress({ start: () => this.handleStepDown(), end: () => this.handleChange() })}
+                      tabindex="-1"
+                    >
+                      <slot name="decrement-number-stepper">
+                        <sd-icon
+                          library="global-resources"
+                          name="system/minus-round"
+                          label="Decrease value"
+                          class=${cx(iconColor, iconMarginLeft, iconSize)}
+                        ></sd-icon>
+                      </slot>
+                    </button>
+
+                    <button
+                      part="increment-number-stepper"
+                      class="stepper-button flex"
+                      type="button"
+                      ?disabled=${this.isIncrementDisabled()}
+                      aria-hidden="true"
+                      ${longPress({ start: () => this.handleStepUp(), end: () => this.handleChange() })}
+                      tabindex="-1"
+                    >
+                      <slot name="increment-number-stepper">
+                        <sd-icon
+                          library="global-resources"
+                          name="system/plus-round"
+                          label="Decrease value"
+                          class=${cx(iconColor, iconMarginLeft, iconSize)}
+                        ></sd-icon>
+                      </slot>
+                    </button>
+                  </div>
+                `
+              : ''}
           </div>
         </div>
 
@@ -713,6 +803,10 @@ export default class SdInput extends SolidElement implements SolidFormControl {
       /* Hides calendar picker for datetime-local type. Does not work in Firefox! */
       input[type='datetime-local']::-webkit-calendar-picker-indicator {
         @apply hidden;
+      }
+
+      .stepper-button[disabled] sd-icon {
+        @apply text-neutral-500;
       }
     `
   ];